pyproject.toml
[tool.poetry]
name = "yolov8"
version = "0.1.0"
description = ""
authors = ["Gurbaj Singh <singh.gurbaj5124871@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ultralytics = "^8.0.190"
pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
